doc: additional document and example gen qr
