--- main.go
@@ -1,21 +1,0 @@
-package vietqr
-
-import (
-	"fmt"
-)
-
-func main() {
-	content := GenerateViQR(RequestGenerateViQR{
-		MerchantAccountInformation: MerchantAccountInformation{
-			AccountNo: "999990335280715",
-		},
-		TransactionAmount: "505000",
-
-		AdditionalDataFieldTemplate: AdditionalDataFieldTemplate{
-			Description: "test noi dungtest noi dungtest noi dungtest noi dung",
-		},
-		Mcc: "5139",
-	})
-
-	fmt.Println("content-main:", content)
-}
